+++ src/Pages/AllArchetypesPage/AllArchetypesPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import styles from './AllArchetypesPage.module.css';
 
-import { ALL_ARCHETYPES, DEV_STRING_PRE, GET_ARCHETYPE_DECKLISTS, GET_CARD, GET_LEGEND_FOR_ARCHETYPE } from '../../Data';
+import { ALL_ARCHETYPES, DEV_STRING_PRE, GET_ARCHETYPE_DECKLISTS, ARCHETYPE_TO_LEGEND_ID } from '../../Data';
 import { LegendImage } from '../../Views';
 
 export class AllArchetypesPage extends React.Component {
@@ -14,13 +14,13 @@ export class AllArchetypesPage extends React.Component {
         return (
             <div className={styles.container}>
                 {ALL_ARCHETYPES.map((archetype) => {
-                    const id = GET_LEGEND_FOR_ARCHETYPE(archetype);
+                    const id = ARCHETYPE_TO_LEGEND_ID(archetype);
                     const numLists = GET_ARCHETYPE_DECKLISTS(archetype).length;
-                    const title = `${archetype}: ${numLists} Decklists (${GET_CARD(id).id})`;
+                    const title = `${archetype}: ${numLists} Decklists (${id})`;
                     return (
                         <div className={styles.linkContainer} key={archetype}>
                             <Link to={`/decklists/riftbound/archetype/${archetype.replaceAll(" ", "-").toLowerCase()}`} style={{color: "var(--text-default)"}}>
-                                <LegendImage id={id} imgTitle={title} size={80} subscript={`${numLists}`} subscriptFontSize={20} />
+                                <LegendImage printId={id} imgTitle={title} size={80} subscript={`${numLists}`} subscriptFontSize={20} />
                             </Link>
                         </div>
                     );